+++ app/controllers/reminder_controller.rb
@@ -3,12 +3,12 @@ class ReminderController < ApplicationController
         if(User.exists?(ide: params[:id]))
             if(params[:content] == "")
                 head :bad_request
-            elsif(params[:hours].to_i < 0 || params[:hours].to_i > 24 || params[:minutes].to_i < 0 || params[:minutes].to_i > 60 || params[:importance].to_i < 0 || params[:importance].to_i > 5)
+            elsif(params[:hours].to_i < 0 || params[:hours].to_i > 24 || params[:minutes].to_i < 0 || params[:minutes].to_i > 60 || params[:importance].to_f < 0 || params[:importance].to_f > 5)
                 head :bad_request
             elsif(params[:month].to_i>12 || params[:month].to_i < 0 || params[:day].to_i > 31 || params[:day].to_i < 0 || params[:year].to_i < 0)
                 head :bad_request
             else
-                Reminder.create(ide: params[:id], content: params[:content], hour: params[:hours].to_i,min: params[:minutes].to_i,year: params[:year].to_i,month: params[:month].to_i,day: params[:month].to_i,importance: params[:importance].to_i)
+                Reminder.create(ide: params[:id], content: params[:content], hour: params[:hours].to_i,min: params[:minutes].to_i,year: params[:year].to_i,month: params[:month].to_i,day: params[:month].to_i,importance: params[:importance].to_f)
                 head :ok
             end
         elsif(!(User.exists?(ide: params[:id])))
