+++ db/migrate/20190620090726_create_reminders.rb
@@ -2,7 +2,7 @@ class CreateReminders < ActiveRecord::Migration[5.2]
   def change
     create_table :reminders do |t|
       t.string :content,null: false, default: ""
-      t.integer :importance,null: false, default: ""
+      t.float :importance,null: false, default: ""
       t.integer :year,null: false, default: ""
       t.integer :month,null: false, default: ""
       t.integer :day,null: false, default: ""
