+++ db/migrate/20190620091332_create_ratings.rb
@@ -4,7 +4,7 @@ class CreateRatings < ActiveRecord::Migration[5.2]
       t.string :what_you_do, null: false, default: ""
       t.string :what_you_good, null: false, default: ""
       t.string :what_you_bad, null: false, default: ""
-      t.integer :importance
+      t.float :importance
       t.string :ide,null: false, default: ""
       t.integer :year,null: false, default: 0
       t.integer :month,null: false, default: 0
